requirements.txt
appdirs==1.4.4
Authlib==0.15.4
black==21.5b2
certifi==2021.5.30
chardet==4.0.0
click==8.0.1
colorama==0.4.4
Flask==2.0.1
Flask-Classful==0.14.2
Flask_Cors==3.0.10
flask-mongoengine==1.0.0
idna==2.10
mypy-extensions==0.4.3
neo4j==4.3.1
openpyxl==3.0.7
pathspec==0.8.1
pytest==6.2.4
python-dateutil==2.8.1
python-dotenv==0.18.0
pyyaml==5.4.1
regex==2021.4.4
requests==2.25.1
toml==0.10.2
urllib3==1.26.5
wheel==0.36.2
flask-mongoengine==1.0.0
flask-login==0.5.0
Flask-Mail==0.9.1